`JenkinsRuleTimeoutTest` was flaky
